fixed document render issues
