+++ src/app/layout.tsx
@@ -1,7 +1,5 @@
 import './globals.css'
 import Header from './header'
-import MainPage from '@/Auth/App'
-
 
 export default function RootLayout({
   children,
@@ -17,7 +15,7 @@ export default function RootLayout({
       <head />
       <body>
         <div>
-          <MainPage />
+          Layout
         </div>
       </body>
     </html>
